Config\Configurator: added method addExtension()
I've added the method addExtension() to allow to add the compiler
extensions in an easy way.

The only argument against this can be the instance of the extension 
should be created only on compile. But extensions should be a 
lightweight cheap objects so this is acceptable solution in my mind.
If the application don't compile this is a side issue anyway.