Fix for row insert to PG Table with renamed seq
I had problem with inserting row to table... I found issue with usage of adsrc in getColumns query. In Postgre doc is written.. The adsrc field is historical, and is best not used, because it does not track outside changes that might affect the representation of the default value. Reverse-compiling the adbin field (with pg_get_expr for example) is a better way to display the default value.... so I changed the query to adbin and all works OK for me.